Duplicate general donation data to Contribution
Same as 98796b89c79736a32d38ca4d3692bc19cc80fa1d, but with a different
name. PostgreSQL failed to apply the previous migration as it was unable
to create an index for one of the fields due to it already existing.
It seems like the issue was caused by DonationLetter previously being called
Donation. Worked around by naming the model Contribution instead.